Begin Simple naming changes to start migration to new lithe
This is basically a staging commit so I have a clean starting point to go back
to if I mess something up. So far I am only changing the naming in 'team.c'
because that is the only file I get compilation errors for at first. All of the
changes made in this commit were simple naming changes since the previous
version of lithe (plus removing the dependance on __x86_64__).  From here I
will start to make the semantical changes.