Create dummy secheduler so contexts can die
At team exit, any lithe contexts which still exist are woken up and allowed to
die instead of being blocked on the barrier so they can be rebound to the next
scheduler that comes in. Since this cleanup code doesn't necessarily run
inside a libgomp scheduler (because the cleanup most likely happens from the
main thread of it's parent scheduler), we need to wrap this cleanup in
a libgomp scheduler and rebind the reamining lithe contexts to it, so that
they get woken up and die while associated with the proper type of scheduler.