No longer special case UPTHREAD to set extended API calls
We only really wanted this so that we would not yield our vcores back
very quickly.  The better way to control this is to allow the library to
make requests itself, but adjust the spin count for how long it spins
before deciding to yield the vcore back to the system.  Right now this
is exposed in an environment variable called UPTHREAD_SPIN_COUNT
(motivated by our LITHE_SPIN_COUNT environment variable). We should
think about adding an extended API call for this instead.